backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Werkzeug==2.2.3
PyMySQL==1.1.0
groq
gunicorn